refactor: unit test
* move end 2 end tests to its own file
* it currently fails, but this is the starting point
* sending one MSH with a batch works, but not two.

[ci skip]
